after pressing the button, it shows the option to undo the changes made
