+++ frontend/src/index.jsx
@@ -5,9 +5,17 @@ import store from "./redux/store";
 import {BrowserRouter} from "react-router-dom";
 import Routers from "./components/routers/Routers.jsx";
 import App from "./components/App.jsx";
-import {ThemeProvider} from "@mui/material";
-import theme from "./components/theme/theme";
+import {ThemeProvider} from "@mui/material/styles";
 
+import { createTheme } from '@mui/material/styles';
+
+const theme = createTheme({
+  palette: {
+    background: {
+      default: '#79786B',
+    },
+  },
+});
 if (module.hot) module.hot.accept();
 
 const root = ReactDOM.createRoot(document.getElementById('app'))
@@ -15,7 +23,9 @@ const root = ReactDOM.createRoot(document.getElementById('app'))
 root.render(
         <Provider store={store}>
             <BrowserRouter>
+                    <ThemeProvider theme={theme}>
                 <App/>
+                    </ThemeProvider>
                 <Routers/>
             </BrowserRouter>
         </Provider>
